% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.h.R
\name{ancova}
\alias{ancova}
\title{ANCOVA}
\usage{
ancova(data, dependent, fixedFactors = NULL, covariates = NULL,
  modelTerms = NULL, contrasts = NULL, descPlotsHAxis = NULL,
  descPlotsSepLines = NULL, descPlotsSepPlots = NULL, postHoc = NULL,
  corrTukey = FALSE, corrScheffe = FALSE, corrBonf = FALSE,
  corrHolm = FALSE, compMain = FALSE, compMainCorr = "none",
  descStats = FALSE, homo = FALSE, qq = FALSE, etaSq = FALSE,
  etaSqP = FALSE, omegaSq = FALSE, plotError = "ci", ciWidth = 95,
  ss = "3")
}
\arguments{
\item{data}{the data as a data frame}

\item{dependent}{a string naming the dependent variable in \code{data}}

\item{fixedFactors}{The fixed factors}

\item{covariates}{The covariates}

\item{modelTerms}{The model terms}

\item{contrasts}{.}

\item{descPlotsHAxis}{The Horizontal axis variable}

\item{descPlotsSepLines}{The Separate lines variable}

\item{descPlotsSepPlots}{The Separate plots variable}

\item{postHoc}{Post Hoc Tests}

\item{corrTukey}{Post Hoc Tests Tukey Correction}

\item{corrScheffe}{Post Hoc Tests Scheffe Correction}

\item{corrBonf}{Post Hoc Tests Bonferroni Correction}

\item{corrHolm}{Post Hoc Tests Holm Correction}

\item{compMain}{Compare main effects}

\item{compMainCorr}{Confidence interval adjustment}

\item{descStats}{Descriptive statistics}

\item{homo}{Display Homogeneity tests}

\item{qq}{Display Q-Q plot of residuals}

\item{etaSq}{Effect size n2}

\item{etaSqP}{Effect size partial n2}

\item{omegaSq}{Effect size w2}

\item{plotError}{Specifies the Error Bar definition as either "Confidence 
interval"(default) or "Standard Error"}

\item{ciWidth}{.}

\item{ss}{Sum of squares}
}
\description{
ANCOVA
}

