% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.h.R
\name{linReg}
\alias{linReg}
\title{Linear Regression}
\usage{
linReg(data, dep, cov)
}
\arguments{
\item{data}{the data as a data frame}

\item{dep}{a string naming the dependent variable}

\item{cov}{a vector of strings naming the covariates}
}
\description{
Linear Regression
}

