% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.b.R
\docType{data}
\name{linRegClass}
\alias{linRegClass}
\title{Returns the names of the factor contrasts as they are
defined by the lm function
Returns the names of the factor contrasts as they are
displayed in the coef table}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
linRegClass
}
\description{
Returns the names of the factor contrasts as they are
defined by the lm function
Returns the names of the factor contrasts as they are
displayed in the coef table
}
\keyword{datasets}
