% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear.h.R
\name{logLinear}
\alias{logLinear}
\title{Log-Linear Regression}
\usage{
logLinear(data, counts, factors = NULL, modelTerms = NULL, est = FALSE,
  ci = FALSE, ciWidth = 95)
}
\arguments{
\item{data}{the data as a data frame}

\item{counts}{The counts}

\item{factors}{The factors}

\item{modelTerms}{The model terms}

\item{est}{Estimates}

\item{ci}{Confidence Intervals}

\item{ciWidth}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$table} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$table$asDF}

\code{as.data.frame(results$table)}
}
\description{
Log-Linear Regression
}
