% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logregbin.b.R
\docType{data}
\name{logRegBinClass}
\alias{logRegBinClass}
\title{Returns the names of the factor contrasts as they are
defined by the glm function
Returns the names of the factor contrasts as they are
displayed in the jmv coef table}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
logRegBinClass
}
\description{
Returns the names of the factor contrasts as they are
defined by the glm function
Returns the names of the factor contrasts as they are
displayed in the jmv coef table
}
\keyword{datasets}
