% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptest2.h.R
\name{propTest2}
\alias{propTest2}
\title{Proportion Test (2 Outcomes)}
\usage{
propTest2(data, vars, areCounts = FALSE, testValue = 0.5,
  hypothesis = "notequal", ci = FALSE, ciWidth = 95)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{a vector of strings naming the variables of interest in 
\code{data}}

\item{areCounts}{\code{TRUE} or \code{FALSE} (default), the variables are 
counts}

\item{testValue}{a number (default: 0.5), the value for the null hypothesis}

\item{hypothesis}{\code{'notequal'} (default), \code{'greater'} or 
\code{'less'}, the alternative hypothesis}

\item{ci}{\code{TRUE} or \code{FALSE} (default), provide confidence 
intervals}

\item{ciWidth}{a number between 50 and 99.9 (default: 95), the confidence 
interval width}
}
\description{
Proportion Test (2 Outcomes)
}

