% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptestn.h.R
\name{propTestN}
\alias{propTestN}
\title{Proportion Test (N Outcomes)}
\usage{
propTestN(data, var, counts, expected, ratio = NULL)
}
\arguments{
\item{data}{the data as a data frame}

\item{var}{a string naming the variable of interest in \code{data}}

\item{counts}{.}

\item{expected}{\code{TRUE} or \code{FALSE} (default), whether expected 
counts should be displayed}

\item{ratio}{a string naming the variable representing counts, or NULL if 
each row represents a single observation}
}
\description{
Proportion Test (N Outcomes)
}

