% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovanp.h.R
\name{anovaNP}
\alias{anovaNP}
\title{One Way ANOVA (Non-parametric)}
\usage{
anovaNP(data, deps, group, pairs = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{deps}{a string naming the dependent variable in \code{data}}

\item{group}{a string naming the grouping or independent variable in 
\code{data}}

\item{pairs}{\code{TRUE} or \code{FALSE} (default), perform pairwise 
comparisons}
}
\description{
One Way ANOVA (Non-parametric)
}
\examples{
\dontrun{
anovaNP(data, deps='len', group='dose')
}
}

