% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00meta.R, R/data.R
\docType{data}
\name{bugs}
\alias{bugs}
\title{Bugs (Ryan, Wilde & Crist, 2013)}
\description{
Bugs (Ryan, Wilde & Crist, 2013)

data sets
}
\references{
\url{http://faculty.kutztown.edu/rryan/RESEARCH/PUBS/Ryan,\%20Wilde,\%20\%26\%20Crist\%202013\%20Web\%20exp\%20vs\%20lab.pdf}
}
\author{
Ryan, Wilde & Crist (2013)
}
\keyword{data}
