% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bugs}
\alias{bugs}
\title{data sets}
\description{
data sets
}
\references{
\url{http://faculty.kutztown.edu/rryan/RESEARCH/PUBS/Ryan,\%20Wilde,\%20\%26\%20Crist\%202013\%20Web\%20exp\%20vs\%20lab.pdf}
}
\author{
Ryan, Wilde & Crist (2013)
}
