% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conttablespaired.h.R
\name{contTablesPaired}
\alias{contTablesPaired}
\title{Paired Samples Contingency Tables}
\usage{
contTablesPaired(data, rows, cols, counts = NULL, pcRow = FALSE,
  pcCol = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{rows}{a string naming the variable to use as the rows in the 
contingency table}

\item{cols}{a string naming the variable to use as the columns in the 
contingency table}

\item{counts}{a string naming the variable to use as counts, or NULL if 
each row represents a single observation}

\item{pcRow}{\code{TRUE} or \code{FALSE} (default), provide row percentages}

\item{pcCol}{\code{TRUE} or \code{FALSE} (default), provide column 
percentages}
}
\description{
McNemar test
}
\examples{
\dontrun{
contTablesPaired(data, rows='x', cols='y')
}
}

