% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptestn.h.R
\name{propTestN}
\alias{propTestN}
\title{Proportion Test (N Outcomes)}
\usage{
propTestN(data, var, counts = NULL, expected = FALSE, ratio = NULL)
}
\arguments{
\item{data}{the data as a data frame}

\item{var}{a string naming the variable of interest in \code{data}}

\item{counts}{a string naming a variable in \code{data} containing counts, 
or NULL if each row represents a single observation}

\item{expected}{\code{TRUE} or \code{FALSE} (default), whether expected 
counts should be displayed}

\item{ratio}{a vector of numbers: the expected proportions}
}
\description{
chi² Goodness of fit
}
\examples{
\dontrun{
propTestN(data, var="outcome", ratio=c(1,1,1))
}
}

