% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.h.R
\name{reliability}
\alias{reliability}
\title{Reliability Analysis}
\usage{
reliability(data, vars, alphaScale = TRUE, omegaScale = FALSE,
  meanScale = FALSE, sdScale = FALSE, corPlot = FALSE,
  alphaItems = FALSE, omegaItems = FALSE, meanItems = FALSE,
  sdItems = FALSE, itemRestCor = FALSE)
}
\arguments{
\item{data}{the data as a data frame}

\item{vars}{a vector of strings naming the variables of interest in 
\code{data}}

\item{alphaScale}{\code{TRUE} (default) or \code{FALSE}, provide Cronbach's 
alpha}

\item{omegaScale}{\code{TRUE} or \code{FALSE} (default), provide McDonald's 
omega}

\item{meanScale}{\code{TRUE} or \code{FALSE} (default), provide the mean}

\item{sdScale}{\code{TRUE} or \code{FALSE} (default), provide the standard 
deviation}

\item{corPlot}{\code{TRUE} or \code{FALSE} (default), provide a correlation 
plot}

\item{alphaItems}{\code{TRUE} or \code{FALSE} (default), provide what the 
Cronbach's alpha would be if the item was dropped}

\item{omegaItems}{\code{TRUE} or \code{FALSE} (default), provide what the 
McDonald's omega would be if the item was dropped}

\item{meanItems}{\code{TRUE} or \code{FALSE} (default), provide item means}

\item{sdItems}{\code{TRUE} or \code{FALSE} (default), provide item standard 
deviations}

\item{itemRestCor}{\code{TRUE} or \code{FALSE} (default), provide item-rest 
correlations}
}
\description{
Reliability Analysis
}
\examples{
\dontrun{
reliability(data, vars=c('x', 'y', 'z'))
}
}
